+++ skbio/stats/evolve/_hommola.py
@@ -10,6 +10,7 @@ import numpy as np
 from scipy.stats import pearsonr
 
 from skbio import DistanceMatrix
+from skbio.stats.distance._base import extract_distance_matrix_data
 from skbio.util import get_rng
 
 
@@ -178,9 +179,12 @@ def hommola_cospeciation(host_dist, par_dist, interaction, permutations=999, see
     pars_k_labels, pars_t_labels = _gen_lists(pars)
     hosts_k_labels, hosts_t_labels = _gen_lists(hosts)
 
+    host_data = extract_distance_matrix_data(host_dist)
+    par_data = extract_distance_matrix_data(par_dist)
+
     # get a vector of pairwise distances for each interaction edge
-    x = _get_dist(hosts_k_labels, hosts_t_labels, host_dist.data, np.arange(num_hosts))
-    y = _get_dist(pars_k_labels, pars_t_labels, par_dist.data, np.arange(num_pars))
+    x = _get_dist(hosts_k_labels, hosts_t_labels, host_data, np.arange(num_hosts))
+    y = _get_dist(pars_k_labels, pars_t_labels, par_data, np.arange(num_pars))
 
     # calculate the observed correlation coefficient for these hosts/symbionts
     corr_coeff = pearsonr(x, y)[0]
@@ -205,8 +209,8 @@ def hommola_cospeciation(host_dist, par_dist, interaction, permutations=999, see
             rng.shuffle(mh)
 
             # get pairwise distances in shuffled order
-            y_p = _get_dist(pars_k_labels, pars_t_labels, par_dist.data, mp)
-            x_p = _get_dist(hosts_k_labels, hosts_t_labels, host_dist.data, mh)
+            y_p = _get_dist(pars_k_labels, pars_t_labels, par_data, mp)
+            x_p = _get_dist(hosts_k_labels, hosts_t_labels, host_data, mh)
 
             # calculate shuffled correlation coefficient
             perm_stats[i] = pearsonr(x_p, y_p)[0]